Refactor nature filter component
