+++ src/BdSp/Underground/Underground.js
@@ -21,11 +21,11 @@ export function Underground() {
     state1: '',
     state2: '',
     state3: '',
-    shiny: 4,
+    shiny_filter: 4,
     min: 0,
     max: 10000,
     delay: 0,
-    nature: [25],
+    nature_filter: [25],
     ability: 3,
     encounter: 12,
     genderRatio: 256,
@@ -58,11 +58,11 @@ export function Underground() {
     state1,
     state2,
     state3,
-    shiny,
+    shiny_filter,
     min,
     max,
     delay,
-    nature,
+    nature_filter,
     ability,
     encounter,
     genderRatio,
@@ -83,11 +83,11 @@ export function Underground() {
       parseInt(state1, 16),
       parseInt(state2, 16),
       parseInt(state3, 16),
-      shiny,
+      shiny_filter,
       min,
       max,
       delay,
-      nature,
+      nature_filter,
       ability,
       encounter,
       genderRatio,
